tutorial_1/requirements.txt
pynacl==1.5.0
pyperclip==1.8.2
git+https://git@github.com/ElrondNetwork/sdk-erdpy-eggs-core.git@init#egg=erdpy_core
git+https://git@github.com/ElrondNetwork/sdk-erdpy-eggs-network-providers.git@init#egg=erdpy_network
git+https://git@github.com/ElrondNetwork/sdk-erdpy-eggs-wallet.git@init#egg=erdpy_wallet
